// src/app/security/shared/user-type.enum.ts
export class UserTypeEnum {

    Doctor: string = "Doctor";
    Nurse: string = "Nurse";
    Receptionist: string = "Receptionist";
    InventoryManager: string = "InventoryManager";
    Cashier: string = "Cashier";
    InsuranceAdmin: string = "InsuranceAdmin";
    SysAdmin: string = "SysAdmin";
    Insurance: string = "Insurance";
    ClinicAdmin: string = "ClinicAdmin";
    Patient: string = "Patient";
    GeneralUser: string = "GeneralUser";
}